% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_beta_ribbon}
\alias{add_beta_ribbon}
\title{Add red "beta" ribbon to the bottom left of the UI}
\usage{
add_beta_ribbon()
}
\value{
html
}
\description{
To be used withi Shiny UI
}
\examples{
htmltools::html_print(add_beta_ribbon())
}
\author{
Joseph Marlo, \email{support@landeranalytics.com}
}
