% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADRCDashHelper-package.R
\docType{package}
\name{ADRCDashHelper-package}
\alias{ADRCDashHelper}
\alias{ADRCDashHelper-package}
\title{ADRCDashHelper: Helper Functions for the ADRC Dashboards}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Chad Murchison \email{AUTHOR@MAIL.COM}

Authors:
\itemize{
  \item Joseph Marlo \email{support@landeranalytics.com}
}

}
\keyword{internal}
