% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{color_palette}
\alias{color_palette}
\title{Color palette}
\format{
An object of class \code{list} of length 8.
}
\usage{
color_palette
}
\value{
list of color hexes
}
\description{
UAB color palette. Primarily used for static UI elements.
}
\details{
Sourced from https://www.uab.edu/toolkit/brand-basics/colors#digital
}
\examples{
color_palette$green
scales::show_col(unlist(color_palette))
}
\author{
Joseph Marlo, \email{support@landeranalytics.com}
}
\keyword{datasets}
