% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_popup_mobile}
\alias{show_popup_mobile}
\title{Show warning popup if user is on mobile}
\usage{
show_popup_mobile(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{A Shiny session object}
}
\value{
called for side effect
}
\description{
To be used within Shiny server
}
\examples{
\dontrun{
observe(if(isTRUE(shinybrowser::is_device_mobile())) ADRCDashHelper::show_popup_mobile(session))
}
}
\author{
Joseph Marlo, \email{support@landeranalytics.com}
}
