% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_uab_bs4_theme}
\alias{write_uab_bs4_theme}
\title{Establish the UAB theme for bs4Dash}
\usage{
write_uab_bs4_theme(output_dir = "inst/app/www")
}
\arguments{
\item{output_dir}{Directory in which to write the style sheet}
}
\value{
Creates or overrides a style sheet at \code{output_dir}
}
\description{
Creates a base UAB theme. It writes out a CSS style sheet to the /www directory. This should be used in the R console when starting a new dashboard project.
}
\examples{
\dontrun{
write_uab_bs4_theme()
}
}
\seealso{
\code{\link[fresh:search_vars_bs4dash]{fresh::search_vars_bs4dash()}}

\url{https://dreamrs.github.io/fresh/articles/vars-bs4dash.html}
}
\author{
Joseph Marlo, \email{support@landeranalytics.com}
}
